MSGraph: added built-in support for pagination
By default, MSGraph returns for a query just the first 100
results ("pagination"). To obtain more results, it is
necessary to issue multiple requests. If "-max_entries" is
specified, the interface tries to get the requested
number of entries. If there are less entries available,
just these are returned.

The parameter -max_entries was added to the following calls:
- group list
- group member list
- user list