Added a ".donecallback" in addition to "-wait" for async operations (via atjobs)
When a "-donecallback" is specified, it is fired when either the async
operation status is "succeeded" or leads to an error. The operation
status is added as additional argument to the callback (it might be a
dict for unexpected replies from MSGraph).

The additional argument was added to the methods
- team archive
- team clone
- team create
- team unarchive